:zap: Finished migrating settings to common and removed batch downloader
