+++ app/src/commonMain/kotlin/com/shub39/rush/di/RushModules.kt
@@ -19,6 +19,7 @@ import com.shub39.rush.lyrics.domain.SongRepo
 import com.shub39.rush.lyrics.presentation.viewmodels.LyricsVM
 import com.shub39.rush.lyrics.presentation.viewmodels.SavedVM
 import com.shub39.rush.lyrics.presentation.viewmodels.SearchSheetVM
+import com.shub39.rush.lyrics.presentation.viewmodels.SettingsVM
 import com.shub39.rush.lyrics.presentation.viewmodels.StateLayer
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.singleOf
@@ -62,4 +63,5 @@ val rushModule = module {
     viewModelOf(::SearchSheetVM)
     viewModelOf(::SavedVM)
     viewModelOf(::LyricsVM)
+    viewModelOf(::SettingsVM)
 }
