+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/AboutLibrariesPage.kt
@@ -2,8 +2,11 @@ package com.shub39.rush.lyrics.presentation.setting
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -12,23 +15,41 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.mikepenz.aboutlibraries.ui.compose.android.rememberLibraries
 import com.mikepenz.aboutlibraries.ui.compose.m3.LibrariesContainer
+import com.mikepenz.aboutlibraries.ui.compose.rememberLibraries
 import com.shub39.rush.core.presentation.PageFill
+import compose.icons.FontAwesomeIcons
+import compose.icons.fontawesomeicons.Solid
+import compose.icons.fontawesomeicons.solid.ArrowLeft
 import org.jetbrains.compose.resources.stringResource
 import rush.app.generated.resources.Res
 import rush.app.generated.resources.about_libraries
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun AboutLibrariesPage() = PageFill {
-    val libraries by rememberLibraries()
+fun AboutLibrariesPage(
+    onNavigateBack: () -> Unit
+) = PageFill {
+    val libraries by rememberLibraries {
+        Res.readBytes("files/aboutlibraries.json").decodeToString()
+    }
 
     Scaffold(
-        modifier = Modifier.widthIn(max = 500.dp),
+        modifier = Modifier.widthIn(max = 1000.dp),
         topBar = {
             TopAppBar(
-                title = { Text(stringResource(Res.string.about_libraries)) }
+                title = { Text(stringResource(Res.string.about_libraries)) },
+                navigationIcon = {
+                    IconButton(
+                        onClick = onNavigateBack
+                    ) {
+                        Icon(
+                            imageVector = FontAwesomeIcons.Solid.ArrowLeft,
+                            contentDescription = "Navigate Back",
+                            modifier = Modifier.size(24.dp)
+                        )
+                    }
+                }
             )
         }
     ) { padding ->
