+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/BackupPage.kt
@@ -12,6 +12,7 @@ import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilledTonalIconButton
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -25,15 +26,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
-import com.shub39.rush.core.domain.data_classes.Theme
 import com.shub39.rush.core.presentation.PageFill
-import com.shub39.rush.core.presentation.RushTheme
 import com.shub39.rush.lyrics.domain.backup.ExportState
 import com.shub39.rush.lyrics.domain.backup.RestoreState
 import compose.icons.FontAwesomeIcons
 import compose.icons.fontawesomeicons.Solid
+import compose.icons.fontawesomeicons.solid.ArrowLeft
 import compose.icons.fontawesomeicons.solid.Check
 import compose.icons.fontawesomeicons.solid.Play
 import io.github.vinceglb.filekit.PlatformFile
@@ -53,7 +52,8 @@ import rush.app.generated.resources.restore_info
 @Composable
 fun BackupPage(
     state: SettingsPageState,
-    action: (SettingsPageAction) -> Unit
+    action: (SettingsPageAction) -> Unit,
+    onNavigateBack: () -> Unit
 ) = PageFill {
     var restoreFile by remember { mutableStateOf<PlatformFile?>(null) }
     val launcher = rememberFilePickerLauncher(
@@ -65,11 +65,22 @@ fun BackupPage(
     }
 
     Scaffold(
-        modifier = Modifier.widthIn(max = 500.dp),
+        modifier = Modifier.widthIn(max = 1000.dp),
         topBar = {
             TopAppBar(
                 title = {
                     Text(stringResource(Res.string.backup))
+                },
+                navigationIcon = {
+                    IconButton(
+                        onClick = onNavigateBack
+                    ) {
+                        Icon(
+                            imageVector = FontAwesomeIcons.Solid.ArrowLeft,
+                            contentDescription = "Navigate Back",
+                            modifier = Modifier.size(24.dp)
+                        )
+                    }
                 }
             )
         }
@@ -170,17 +181,5 @@ fun BackupPage(
                 }
             )
         }
-    }
-}
-
-@PreviewLightDark
-@Composable
-private fun Preview() {
-    RushTheme(
-        state = Theme()
-    ) {
-        BackupPage(
-            SettingsPageState()
-        ) { }
     }
 }
