+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/LookAndFeelPage.kt
@@ -1,55 +1,31 @@
 package com.shub39.rush.lyrics.presentation.setting
 
-import android.R.color.system_accent1_200
-import android.os.Build
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Create
-import androidx.compose.material.icons.outlined.Check
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilledTonalIconButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.ListItem
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.PlainTooltip
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.TooltipBox
-import androidx.compose.material3.TooltipDefaults
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.contentColorFor
-import androidx.compose.material3.rememberTooltipState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -57,26 +33,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.materialkolor.PaletteStyle
-import com.materialkolor.ktx.from
-import com.materialkolor.palettes.TonalPalette
-import com.materialkolor.rememberDynamicColorScheme
-import com.shub39.rush.core.domain.data_classes.Theme
 import com.shub39.rush.core.domain.enums.AppTheme
 import com.shub39.rush.core.domain.enums.Fonts
 import com.shub39.rush.core.presentation.ColorPickerDialog
 import com.shub39.rush.core.presentation.PageFill
 import com.shub39.rush.core.presentation.RushDialog
-import com.shub39.rush.core.presentation.RushTheme
 import compose.icons.FontAwesomeIcons
 import compose.icons.fontawesomeicons.Solid
+import compose.icons.fontawesomeicons.solid.ArrowLeft
 import compose.icons.fontawesomeicons.solid.FolderOpen
 import org.jetbrains.compose.resources.Font
 import org.jetbrains.compose.resources.stringResource
@@ -85,9 +53,6 @@ import rush.app.generated.resources.amoled
 import rush.app.generated.resources.amoled_desc
 import rush.app.generated.resources.font
 import rush.app.generated.resources.look_and_feel
-import rush.app.generated.resources.material_theme
-import rush.app.generated.resources.material_theme_desc
-import rush.app.generated.resources.palette_style
 import rush.app.generated.resources.seed_color
 import rush.app.generated.resources.seed_color_desc
 import rush.app.generated.resources.select_app_theme
@@ -96,20 +61,32 @@ import rush.app.generated.resources.select_app_theme
 @Composable
 fun LookAndFeelPage(
     state: SettingsPageState,
-    action: (SettingsPageAction) -> Unit
+    action: (SettingsPageAction) -> Unit,
+    onNavigateBack: () -> Unit
 ) = PageFill {
     var colorPickerDialog by remember { mutableStateOf(false) }
     var fontPickerDialog by remember { mutableStateOf(false) }
     var themePickerDialog by remember { mutableStateOf(false) }
 
     Scaffold(
-        modifier = Modifier.widthIn(max = 500.dp),
+        modifier = Modifier.widthIn(max = 1000.dp),
         topBar = {
             TopAppBar(
                 title = {
                     Text(
                         text = stringResource(Res.string.look_and_feel)
                     )
+                },
+                navigationIcon = {
+                    IconButton(
+                        onClick = onNavigateBack
+                    ) {
+                        Icon(
+                            imageVector = FontAwesomeIcons.Solid.ArrowLeft,
+                            contentDescription = "Navigate Back",
+                            modifier = Modifier.size(24.dp)
+                        )
+                    }
                 }
             )
         }
@@ -196,32 +173,7 @@ fun LookAndFeelPage(
             }
 
             // Material you toggle
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-                item {
-                    ListItem(
-                        headlineContent = {
-                            Text(
-                                text = stringResource(Res.string.material_theme)
-                            )
-                        },
-                        supportingContent = {
-                            Text(
-                                text = stringResource(Res.string.material_theme_desc)
-                            )
-                        },
-                        trailingContent = {
-                            Switch(
-                                checked = state.theme.materialTheme,
-                                onCheckedChange = {
-                                    action(
-                                        SettingsPageAction.OnMaterialThemeToggle(it)
-                                    )
-                                }
-                            )
-                        }
-                    )
-                }
-            }
+            materialYouToggle(state, action)
 
             // Seed color picker
             item {
@@ -255,57 +207,7 @@ fun LookAndFeelPage(
             }
 
             // palette styles
-            item {
-                ListItem(
-                    headlineContent = {
-                        Text(
-                            text = stringResource(Res.string.palette_style)
-                        )
-                    },
-                    supportingContent = {
-                        val scrollState = rememberScrollState()
-
-                        Row(
-                            modifier = Modifier
-                                .horizontalScroll(scrollState)
-                                .padding(vertical = 8.dp),
-                            horizontalArrangement = Arrangement.spacedBy(8.dp),
-                        ) {
-                            PaletteStyle.entries.toList().forEach { style ->
-                                val scheme = rememberDynamicColorScheme(
-                                    primary = if (state.theme.materialTheme && Build.VERSION.SDK_INT > Build.VERSION_CODES.S) {
-                                        colorResource(system_accent1_200)
-                                    } else {
-                                        Color(state.theme.seedColor)
-                                    },
-                                    isDark = when (state.theme.appTheme) {
-                                        AppTheme.SYSTEM -> isSystemInDarkTheme()
-                                        AppTheme.LIGHT -> false
-                                        AppTheme.DARK -> true
-                                    },
-                                    isAmoled = state.theme.withAmoled,
-                                    style = style
-                                )
-
-                                SelectableMiniPalette(
-                                    selected = state.theme.style == style,
-                                    onClick = {
-                                        action(
-                                            SettingsPageAction.OnPaletteChange(style = style)
-                                        )
-                                    },
-                                    contentDescription = { style.name },
-                                    accents = listOf(
-                                        TonalPalette.from(scheme.primary),
-                                        TonalPalette.from(scheme.tertiary),
-                                        TonalPalette.from(scheme.secondary)
-                                    )
-                                )
-                            }
-                        }
-                    }
-                )
-            }
+            paletteStyles(state, action)
 
             item {
                 Spacer(modifier = Modifier.padding(60.dp))
@@ -398,91 +300,5 @@ fun LookAndFeelPage(
                 }
             }
         }
-    }
-}
-
-// yeeted from https://github.com/SkyD666/PodAura/blob/master/app/src/main/java/com/skyd/anivu/ui/screen/settings/appearance/AppearanceScreen.kt
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun SelectableMiniPalette(
-    modifier: Modifier = Modifier,
-    selected: Boolean,
-    onClick: () -> Unit,
-    contentDescription: () -> String,
-    accents: List<TonalPalette>,
-) {
-    Surface(
-        modifier = modifier,
-        shape = RoundedCornerShape(16.dp),
-        color = MaterialTheme.colorScheme.inverseOnSurface,
-    ) {
-        TooltipBox(
-            modifier = modifier,
-            positionProvider = TooltipDefaults.rememberPlainTooltipPositionProvider(),
-            tooltip = {
-                PlainTooltip {
-                    Text(contentDescription())
-                }
-            },
-            state = rememberTooltipState()
-        ) {
-            Surface(
-                modifier = Modifier
-                    .clickable(onClick = onClick)
-                    .padding(12.dp)
-                    .size(50.dp),
-                shape = CircleShape,
-                color = Color(accents[0].tone(60))
-            ) {
-                Box {
-                    Surface(
-                        modifier = Modifier
-                            .size(50.dp)
-                            .offset((-25).dp, 25.dp),
-                        color = Color(accents[1].tone(85)),
-                    ) {}
-                    Surface(
-                        modifier = Modifier
-                            .size(50.dp)
-                            .offset(25.dp, 25.dp),
-                        color = Color(accents[2].tone(75)),
-                    ) {}
-                    val animationSpec = spring<Float>(stiffness = Spring.StiffnessMedium)
-                    AnimatedVisibility(
-                        visible = selected,
-                        enter = scaleIn(animationSpec) + fadeIn(animationSpec),
-                        exit = scaleOut(animationSpec) + fadeOut(animationSpec),
-                    ) {
-                        Box(
-                            modifier = Modifier
-                                .padding(10.dp)
-                                .fillMaxSize()
-                                .clip(CircleShape)
-                                .background(MaterialTheme.colorScheme.primary),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Icon(
-                                imageVector = Icons.Outlined.Check,
-                                contentDescription = "Checked",
-                                modifier = Modifier
-                                    .padding(8.dp)
-                                    .size(16.dp),
-                                tint = MaterialTheme.colorScheme.surface
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun Preview() {
-    RushTheme(
-        state = Theme()
-    ) {
-        LookAndFeelPage(SettingsPageState()) {}
     }
 }
