+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/SettingRootPage.kt
@@ -1,6 +1,5 @@
 package com.shub39.rush.lyrics.presentation.setting
 
-import android.content.Intent
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -11,12 +10,12 @@ import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.Warning
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilledTonalIconButton
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
@@ -29,31 +28,24 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import com.shub39.rush.core.presentation.PageFill
 import com.shub39.rush.core.presentation.RushDialog
-import com.shub39.rush.core.presentation.RushTheme
-import com.shub39.rush.lyrics.presentation.SettingsRoutes
 import compose.icons.FontAwesomeIcons
 import compose.icons.fontawesomeicons.Solid
+import compose.icons.fontawesomeicons.solid.ArrowLeft
 import compose.icons.fontawesomeicons.solid.ArrowRight
 import org.jetbrains.compose.resources.stringResource
 import rush.app.generated.resources.Res
 import rush.app.generated.resources.about_libraries
 import rush.app.generated.resources.backup
 import rush.app.generated.resources.backup_info
-import rush.app.generated.resources.batch_download
-import rush.app.generated.resources.batch_download_info
 import rush.app.generated.resources.delete_all
 import rush.app.generated.resources.delete_confirmation
-import rush.app.generated.resources.grant_permission
 import rush.app.generated.resources.look_and_feel
 import rush.app.generated.resources.look_and_feel_info
-import rush.app.generated.resources.notification_permission
 import rush.app.generated.resources.settings
 
 // topmost settings page
@@ -62,19 +54,29 @@ import rush.app.generated.resources.settings
 fun SettingRootPage(
     notificationAccess: Boolean,
     action: (SettingsPageAction) -> Unit,
-    navigator: (SettingsRoutes) -> Unit
+    navigator: (SettingsRoutes) -> Unit,
+    onNavigateBack: () -> Unit
 ) = PageFill {
-    val context = LocalContext.current
-
     var deleteButtonStatus by remember { mutableStateOf(true) }
     var deleteConfirmationDialog by remember { mutableStateOf(false) }
 
     Scaffold(
-        modifier = Modifier.widthIn(max = 500.dp),
+        modifier = Modifier.widthIn(max = 1000.dp),
         topBar = {
             TopAppBar(
                 title = {
                     Text(stringResource(Res.string.settings))
+                },
+                navigationIcon = {
+                    IconButton(
+                        onClick = onNavigateBack
+                    ) {
+                        Icon(
+                            imageVector = FontAwesomeIcons.Solid.ArrowLeft,
+                            contentDescription = "Navigate Back",
+                            modifier = Modifier.size(24.dp)
+                        )
+                    }
                 }
             )
         }
@@ -121,25 +123,6 @@ fun SettingRootPage(
                 )
             }
 
-            // navigate to batch downloader
-            item {
-                ListItem(
-                    headlineContent = { Text(text = stringResource(Res.string.batch_download)) },
-                    supportingContent = { Text(text = stringResource(Res.string.batch_download_info)) },
-                    trailingContent = {
-                        FilledTonalIconButton(
-                            onClick = { navigator(SettingsRoutes.BatchDownloaderPage) },
-                        ) {
-                            Icon(
-                                imageVector = FontAwesomeIcons.Solid.ArrowRight,
-                                contentDescription = "Navigate",
-                                modifier = Modifier.size(20.dp)
-                            )
-                        }
-                    }
-                )
-            }
-
             // navigate to backup
             item {
                 ListItem(
@@ -160,26 +143,7 @@ fun SettingRootPage(
             }
 
             // navigate to notification access permission page
-            if (!notificationAccess) {
-                item {
-                    val intent = Intent("android.settings.ACTION_NOTIFICATION_LISTENER_SETTINGS")
-
-                    ListItem(
-                        headlineContent = { Text(text = stringResource(Res.string.grant_permission)) },
-                        supportingContent = { Text(text = stringResource(Res.string.notification_permission)) },
-                        trailingContent = {
-                            FilledTonalIconButton(
-                                onClick = { context.startActivity(intent) }
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.Info,
-                                    contentDescription = null
-                                )
-                            }
-                        }
-                    )
-                }
-            }
+            notificationAccessReminder(notificationAccess)
 
             // navigate to about app
             item {
@@ -249,17 +213,5 @@ fun SettingRootPage(
                 }
             }
         }
-    }
-}
-
-@PreviewLightDark
-@Composable
-private fun Preview() {
-    RushTheme {
-        SettingRootPage(
-            notificationAccess = false,
-            action = {},
-            navigator = {}
-        )
     }
 }
