+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/SettingsRoutes.kt
@@ -1,0 +1,17 @@
+package com.shub39.rush.lyrics.presentation.setting
+
+import kotlinx.serialization.Serializable
+
+sealed interface SettingsRoutes {
+    @Serializable
+    data object SettingRootPage : SettingsRoutes
+
+    @Serializable
+    data object BackupPage : SettingsRoutes
+
+    @Serializable
+    data object LookAndFeelPage : SettingsRoutes
+
+    @Serializable
+    data object AboutLibrariesPage : SettingsRoutes
+}
