+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/component/SelectableMiniPalette.kt
@@ -1,0 +1,112 @@
+package com.shub39.rush.lyrics.presentation.setting.component
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.Spring
+import androidx.compose.animation.core.spring
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.scaleIn
+import androidx.compose.animation.scaleOut
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Check
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.PlainTooltip
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TooltipBox
+import androidx.compose.material3.TooltipDefaults
+import androidx.compose.material3.rememberTooltipState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.materialkolor.palettes.TonalPalette
+
+// yeeted from https://github.com/SkyD666/PodAura/blob/master/app/src/main/java/com/skyd/anivu/ui/screen/settings/appearance/AppearanceScreen.kt
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SelectableMiniPalette(
+    modifier: Modifier = Modifier,
+    selected: Boolean,
+    onClick: () -> Unit,
+    contentDescription: () -> String,
+    accents: List<TonalPalette>,
+) {
+    Surface(
+        modifier = modifier,
+        shape = RoundedCornerShape(16.dp),
+        color = MaterialTheme.colorScheme.inverseOnSurface,
+    ) {
+        TooltipBox(
+            modifier = modifier,
+            positionProvider = TooltipDefaults.rememberPlainTooltipPositionProvider(),
+            tooltip = {
+                PlainTooltip {
+                    Text(contentDescription())
+                }
+            },
+            state = rememberTooltipState()
+        ) {
+            Surface(
+                modifier = Modifier
+                    .clickable(onClick = onClick)
+                    .padding(12.dp)
+                    .size(50.dp),
+                shape = CircleShape,
+                color = Color(accents[0].tone(60))
+            ) {
+                Box {
+                    Surface(
+                        modifier = Modifier
+                            .size(50.dp)
+                            .offset((-25).dp, 25.dp),
+                        color = Color(accents[1].tone(85)),
+                    ) {}
+                    Surface(
+                        modifier = Modifier
+                            .size(50.dp)
+                            .offset(25.dp, 25.dp),
+                        color = Color(accents[2].tone(75)),
+                    ) {}
+                    val animationSpec = spring<Float>(stiffness = Spring.StiffnessMedium)
+                    AnimatedVisibility(
+                        visible = selected,
+                        enter = scaleIn(animationSpec) + fadeIn(animationSpec),
+                        exit = scaleOut(animationSpec) + fadeOut(animationSpec),
+                    ) {
+                        Box(
+                            modifier = Modifier
+                                .padding(10.dp)
+                                .fillMaxSize()
+                                .clip(CircleShape)
+                                .background(MaterialTheme.colorScheme.primary),
+                            contentAlignment = Alignment.Center
+                        ) {
+                            Icon(
+                                imageVector = Icons.Outlined.Check,
+                                contentDescription = "Checked",
+                                modifier = Modifier
+                                    .padding(8.dp)
+                                    .size(16.dp),
+                                tint = MaterialTheme.colorScheme.surface
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
