+++ app/src/commonMain/kotlin/com/shub39/rush/lyrics/presentation/setting/util.kt
@@ -1,0 +1,15 @@
+package com.shub39.rush.lyrics.presentation.setting
+
+import androidx.compose.foundation.lazy.LazyListScope
+
+expect fun LazyListScope.materialYouToggle(
+    state: SettingsPageState,
+    action: (SettingsPageAction) -> Unit
+)
+
+expect fun LazyListScope.paletteStyles(
+    state: SettingsPageState,
+    action: (SettingsPageAction) -> Unit
+)
+
+expect fun LazyListScope.notificationAccessReminder(notificationAccess: Boolean)
