--- app/src/desktopMain/kotlin/com/shub39/rush/lyrics/presentation/SettingsGraph.kt
@@ -1,16 +1,0 @@
-package com.shub39.rush.lyrics.presentation
-
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun SettingsGraph(
-    onBack: () -> Unit,
-) {
-    Button(
-        onClick = onBack
-    ) {
-        Text("WIP")
-    }
-}
